// client/app/po/purchase-order.ts
export class PurchaseOrder {
    docno: number;
    doctype: string;
    date: Date;
    vendor: string;
    porg: string;
    htext: string;
    delivery: string;
    terms: string;
    termstext1: string;
    termstext2: string;
    termstext3: string;
    createdby: string;
    createstamp: Date;
}